[8.x] Add tests for `LazyCollection` constructor laziness (#36846)
* Move enumeration counting to trait

* Add tests for `LazyCollection` constructor laziness
